mtime tests, small tearDown change

+++ test.py
@@ -4,8 +4,10 @@ from mutagen.mp3 import EasyMP3 as MP3
 from flask import Flask
 from flask.ext.testing import TestCase
 import unittest
+import time
 from models import db, Track
 from manage import populate_db, update_db
+
 
 class MyTest(TestCase):
 
@@ -33,7 +35,7 @@ def create_mock_tracks(tracks, src_track="test/sinewave.mp3"):
         filename = 'test/' + track
         shutil.copyfile(src_track, filename)
         song_tag = MP3(filename)
-        for k,v in tracks[track].iteritems():
+        for k, v in tracks[track].iteritems():
             song_tag[k] = unicode(v)
         song_tag.save()
 
@@ -80,17 +82,19 @@ class UpdateTest(MyTest):
         create_mock_tracks(self.mock_tracks)
 
     def test_added_track_update(self):
+        ''' db is updated to new files in music_dir '''
         added_track = self.added_track
         populate_db('test', False)
         create_mock_tracks(added_track)
         update_db('test', False)
         filename = added_track.keys()[0]
-        found_track = Track.query.filter_by(artist=added_track[filename]['artist'],
-                                         title=added_track[filename]['title'])
+        found_track = Track.query.filter_by(
+                                    artist=added_track[filename]['artist'],
+                                    title=added_track[filename]['title'])
         assert found_track is not None
 
     def test_remove_track_update(self):
-        # populate first, then alter and update
+        ''' db doesn't included deleted tracks '''
         mock_tracks = self.mock_tracks
         # pop returns a tuple of (filename, file_info_dict)
         removed_track = mock_tracks.popitem()[0]  # just track name
@@ -99,9 +103,26 @@ class UpdateTest(MyTest):
         tracks_in_db = Track.query.all()
         assert removed_track not in tracks_in_db
 
+    def test_mtime(self):
+        ''' newest mtime is updated in db '''
+        mock_tracks = self.mock_tracks
+        now = int(time.time())
+        for f in mock_tracks:
+            filename = os.path.join('test', f)
+            st = os.stat(filename)
+            atime = st.st_atime
+            new_mtime = now
+            os.utime(filename, (atime, new_mtime))  # modify the timestamp
+        update_db('test', False)
+        tracks_in_db = Track.query.all()
+        for track in tracks_in_db:
+            assert track.mtime == now
+
     def tearDown(self):
-        for track in self.mock_tracks:
-            os.remove("test/" + track)
+        ''' remove all but sinewave.mp3 '''
+        for track in os.listdir('test'):
+            if track != 'sinewave.mp3':
+                os.remove("test/" + track)
         super(UpdateTest, self).tearDown()
